+++ src/main/java/persistence/sql/ddl/converter/TypeConverter.java
@@ -1,4 +1,4 @@
-package persistence.sql.ddl;
+package persistence.sql.ddl.converter;
 
 public interface TypeConverter {
