+++ src/test/java/persistence/sql/ddl/H2TypeTest.java
@@ -4,6 +4,7 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.MethodSource;
+import persistence.sql.ddl.model.H2Type;
 
 import java.util.stream.Stream;
